WIP on making the test menu
